Finalizar separación de cleverbot del script principal
